+++ src/resource.ts
@@ -1,8 +1,9 @@
 import Client from './client';
-import { State, headStateFactory, HeadState } from './state';
+import { State, headStateFactory, HeadState, isState } from './state';
 import { resolve } from './util/uri';
 import { FollowPromiseOne, FollowPromiseMany } from './follow-promise';
 import { Link, LinkNotFound, LinkVariables } from './link';
+import { EventEmitter } from 'events';
 import { GetRequestOptions, PostRequestOptions, PatchRequestOptions, PutRequestOptions, HeadRequestOptions } from './types';
 
 /**
@@ -14,7 +15,7 @@ import { GetRequestOptions, PostRequestOptions, PatchRequestOptions, PutRequestO
  * A resource may also have a list of links on them, pointing to other
  * resources.
  */
-export default class Resource<T = any> {
+export class Resource<T = any> extends EventEmitter {
 
   uri: string;
   client: Client;
@@ -25,6 +26,7 @@ export default class Resource<T = any> {
    * uri must be absolute
    */
   constructor(client: Client, uri: string) {
+    super();
     this.client = client;
     this.uri = uri;
     this.activeRefresh = null;
@@ -86,7 +88,7 @@ export default class Resource<T = any> {
         try {
           const response = await this.fetchOrThrow(params);
           const state = await this.client.getStateForResponse(this.uri, response);
-          this.client.cache.store(state);
+          this.client.cacheState(state);
           return state;
         } finally {
           this.activeRefresh = null;
@@ -101,11 +103,29 @@ export default class Resource<T = any> {
   /**
    * Updates the server state with a PUT request
    */
-  async put(options: PutRequestOptions<T>): Promise<void> {
+  async put(options: PutRequestOptions<T> | State): Promise<void> {
 
-    await this.fetchOrThrow(
-      optionsToRequestInit('PUT', options)
-    );
+
+    const requestInit = optionsToRequestInit('PUT', options);
+
+    /**
+     * If we got a 'State' object passed, it means we don't need to emit a
+     * stale event, as the passed object is the new
+     * state.
+     *
+     * We're gonna track that with a custom header that will be removed
+     * later in the fetch pipeline.
+     */
+    if (isState(options)) {
+      requestInit.headers.set('X-KETTING-NO-STALE', '1');
+    }
+
+    await this.fetchOrThrow(requestInit);
+
+    if (isState(options)) {
+      this.client.cache.store(options);
+      this.emit('update', options);
+    }
 
   }
 
@@ -301,21 +321,48 @@ export default class Resource<T = any> {
   }
 }
 
+export declare interface Resource<T = any> {
+
+  on(event: 'update', listener: (state: State) => void) : this
+  on(event: 'stale',  listener: () => void) : this;
+  on(event: 'delete', listener: () => void) : this;
+
+  once(event: 'update', listener: (state: State) => void) : this
+  once(event: 'stale',  listener: () => void) : this;
+  once(event: 'delete', listener: () => void) : this;
+
+  off(event: 'update', listener: (state: State) => void) : this
+  off(event: 'stale',  listener: () => void) : this;
+  off(event: 'delete', listener: () => void) : this;
+
+  emit(event: 'update', state: State) : boolean
+  emit(event: 'stale') : boolean;
+  emit(event: 'delete') : boolean;
+
+}
+
+export default Resource;
+
+type StrictRequestInit = RequestInit & {
+  headers: Headers,
+};
+
 /**
  * Convert request options to RequestInit
  *
  * RequestInit is passed to the constructor of fetch(). We have our own 'options' format
  */
-function optionsToRequestInit(method: 'GET', options?: GetRequestOptions): RequestInit;
-function optionsToRequestInit(method: 'HEAD', options?: HeadRequestOptions): RequestInit;
-function optionsToRequestInit(method: 'PATCH', options?: PatchRequestOptions): RequestInit;
-function optionsToRequestInit(method: 'POST', options?: PostRequestOptions): RequestInit;
-function optionsToRequestInit(method: 'PUT', options?: PutRequestOptions): RequestInit;
-function optionsToRequestInit(method: string, options?: GetRequestOptions | PostRequestOptions | PatchRequestOptions | PutRequestOptions): RequestInit {
+function optionsToRequestInit(method: 'GET', options?: GetRequestOptions): StrictRequestInit;
+function optionsToRequestInit(method: 'HEAD', options?: HeadRequestOptions): StrictRequestInit;
+function optionsToRequestInit(method: 'PATCH', options?: PatchRequestOptions): StrictRequestInit;
+function optionsToRequestInit(method: 'POST', options?: PostRequestOptions): StrictRequestInit;
+function optionsToRequestInit(method: 'PUT', options?: PutRequestOptions): StrictRequestInit;
+function optionsToRequestInit(method: string, options?: GetRequestOptions | PostRequestOptions | PatchRequestOptions | PutRequestOptions): StrictRequestInit {
 
   if (!options) {
     return {
-      method
+      method,
+      headers: new Headers(),
     };
   }
   let headers;
@@ -343,7 +390,7 @@ function optionsToRequestInit(method: string, options?: GetRequestOptions | Post
   return {
     method,
     body,
-    headers
+    headers,
   };
 
 }