+++ src/state/binary.ts
@@ -1,6 +1,7 @@
 import { BaseState } from './base-state';
 import { StateFactory } from './interface';
 import { parseLink } from '../http/util';
+import { Links } from '../link';
 
 /**
  * Represents a binary resource state.
@@ -12,6 +13,17 @@ export class BinaryState extends BaseState<Blob> {
   serializeBody(): Blob {
 
     return this.data;
+
+  }
+
+  clone(): BinaryState {
+
+    return new BinaryState(
+      this.uri,
+      this.data,
+      new Headers(this.headers),
+      new Links(this.links)
+    );
 
   }