+++ src/state/html.ts
@@ -2,6 +2,7 @@ import { BaseState } from './base-state';
 import { StateFactory } from './interface';
 import { parseLink } from '../http/util';
 import { parseHtmlLinks } from '../util/html';
+import { Links } from '../link';
 
 /**
  * Represents a resource state in the HAL format
@@ -11,6 +12,17 @@ export class HtmlState extends BaseState<string> {
   serializeBody(): string {
 
     throw new Error('Reserializing HTML states is not yet supported. Please log an issue in the Ketting project to help figure out how this should be done');
+
+  }
+
+  clone(): HtmlState {
+
+    return new HtmlState(
+      this.uri,
+      this.data,
+      new Headers(this.headers),
+      new Links(this.links)
+    );
 
   }