+++ src/state/text.ts
@@ -1,6 +1,7 @@
 import { BaseState } from './base-state';
 import { StateFactory } from './interface';
 import { parseLink } from '../http/util';
+import { Links } from '../link';
 
 /**
  * Represents a resource state for text responses, such as text/plain, text/csv.
@@ -11,6 +12,17 @@ export class TextState extends BaseState<string> {
   serializeBody(): string {
 
     return this.data;
+
+  }
+
+  clone(): TextState {
+
+    return new TextState(
+      this.uri,
+      this.data,
+      new Headers(this.headers),
+      new Links(this.links),
+    );
 
   }