+++ test/unit/client.ts
@@ -5,7 +5,7 @@ describe('Client', () => {
 
   describe('Resource caching', () => {
 
-    it('should invalidate a resource\'s cache if an unsafe method was used', () => {
+    it('should invalidate a resource\'s cache if an unsafe method was used', async () => {
 
       const client = new Client('https://example.org');
       client.cache.store(new TextState(
@@ -15,13 +15,14 @@ describe('Client', () => {
         new Links(),
       ));
 
+      client.use( async req => {
+        return new Response('OK');
+      });
       const request = new Request('https://example.org/foo', {
         method: 'POST'
       });
 
-      const response = new Response('', {status: 200});
-      client.cache.processRequest(request, response);
-
+      await client.fetcher.fetch(request);
       expect(client.cache.has('https://example.org/foo')).to.equal(false);
 
     });
@@ -40,14 +41,16 @@ describe('Client', () => {
         method: 'SEARCH'
       });
 
-      const response = new Response('', {status: 200});
-      client.cache.processRequest(request, response);
+      client.use( async req => {
+        return new Response('OK');
+      });
 
+      client.fetcher.fetch(request);
       expect(client.cache.has('https://example.org/foo')).to.equal(true);
 
     });
 
-    it('should invalidate resources if they were mentioned in a Link header with rel="invalidates"', () => {
+    it('should invalidate resources if they were mentioned in a Link header with rel="invalidates"', async () => {
 
       const client = new Client('https://example.org');
       client.cache.store(new TextState(
@@ -64,12 +67,11 @@ describe('Client', () => {
       const headers = new Headers();
       headers.append('Link', '</bar>; rel="invalidates"');
       headers.append('Link', '</zim>; rel="invalidates"');
-      const response = new Response('', {
-        status: 200,
-        headers
+      client.use( async req => {
+        return new Response('OK', { headers });
       });
 
-      client.cache.processRequest(request, response);
+      await client.fetcher.fetch(request);
       expect(client.cache.has('https://example.org/foo')).to.equal(false);
 
     });
@@ -93,6 +95,7 @@ describe('Client', () => {
       });
 
       const halState = await client.getStateForResponse('https://example.org/parent', response);
+      client.cacheState(halState);
       expect(halState.uri).to.equal('https://example.org/parent');
       expect(halState.data).to.eql({item: 2});
 
@@ -132,6 +135,7 @@ describe('Client', () => {
       });
 
       const halState = await client.getStateForResponse('https://example.org/parent', response);
+      client.cacheState(halState);
       expect(halState.uri).to.equal('https://example.org/parent');
       expect(halState.data).to.eql({item: 2});